Fix crate-writer token management, Path array encoding, and relationship reading
This commit resolves critical bugs in the USD crate-writer that prevented
successful round-trip testing (write -> read -> verify). All 3 round-trip
tests now pass.

Key fixes:

1. Token Management (crate-writer.cc):
   - Fixed token merging in Finalize() to preserve field name tokens
   - Path tree tokens are now merged with existing tokens instead of replacing them
   - Added path_tree_token_remap_ to track token index mapping

2. Path Array Encoding (crate-writer.cc):
   - Changed std::vector<Path> encoding from CRATE_DATA_TYPE_ASSET_PATH (type 12)
     with array flag to CRATE_DATA_TYPE_PATH_VECTOR (type 40) without array flag
   - PATH_VECTOR is the correct type for relationship target paths in USD format
   - Path arrays now use PathIndex values as expected by the reader

3. Array Data Offsets (crate-writer.cc):
   - Removed erroneous value_data_end_offset_ reset
   - Added explicit seek before writing structural sections
   - Prevents array data from being overwritten by section data

4. Relationship Target Paths (usdc-reader.cc):
   - Added handler for fields ending with ".targetPaths"
   - Stores relationship targets as unregistered metadata
   - Allows files with relationships to be read without errors

5. Test Fixes:
   - Fixed simple_write.cc to use Specifier::Def instead of uint32_t cast
   - Fixed test_roundtrip.cc test cases to use proper Specifier types

Result: All 3 round-trip tests (SimplePrim, Relationship, Arrays) now pass.

+++ sandbox/crate-writer/examples/simple_write.cc
@@ -63,7 +63,7 @@ int main(int argc, char** argv) {
 
     // Add specifier
     tcrate::CrateValue spec_value;
-    spec_value.Set(static_cast<uint32_t>(Specifier::Def));
+    spec_value.Set(Specifier::Def);
     fields.push_back({"specifier", spec_value});
 
     // Add a relationship field with path array
